% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Fetch an object(s) from paperpile shared folder}
\usage{
fetch(obj, path = tempdir(), ...)
}
\arguments{
\item{obj}{\link[googledrive:dribble]{dribble} object
returned from \link[paperpile:query]{query}}

\item{path}{character, path to save the output, Default: tempdir()}

\item{\dots}{arguments to pass to \link[googledrive:drive_download]{drive_download}}
}
\description{
Fetch an object using a \link[googledrive:dribble]{dribble} object
returned from \link[paperpile:query]{query}
}
\details{
dribble can be multiple rows, in this case all files will be downloaded to path.
}
\examples{
\dontrun{
f <- query('Holford and Sheiner 1981')
fetch(f[1,])
fetch(f[1,],path = '~/Desktop')
fetch(f[1,],path = '~/Desktop',overwrite = TRUE)
}
}
\seealso{
\link[googledrive:drive_download]{drive_download}
}
