% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query paperpile google drive}
\usage{
query(
  pattern,
  field = "name",
  operator = "contains",
  type = "application/pdf",
  ...
)
}
\arguments{
\item{pattern}{text to search for in title or in body of document}

\item{field}{character, what field to query, usually name of fullText, Default: 'name'}

\item{operator}{character, what operator to apply in query, usually '=' or 'contains', Default: 'contains'}

\item{type}{character, mime type to search for, Default: 'application/pdf'}

\item{\dots}{arguments passed to \link[googledrive:drive_find]{drive_find}}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Using googledrive api query the paperpile shared folder
}
\details{
DETAILS
}
\examples{
\dontrun{
query('understanding the dose-effect')
}
}
\seealso{
\link[googledrive:drive_find]{drive_find}
}
